Initial conversion to JS styles
